refactor: move entity classes into separate package
